refactor: Split GPSController to saveGPSController and getSummaryController
